% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{bargraphPlot}
\alias{bargraphPlot}
\title{Produces a bar graph}
\usage{
bargraphPlot(virginiaYears, yaxis, yaxisLabel, title)
}
\arguments{
\item{virginiaYears}{data}

\item{yaxis}{variable.}

\item{yaxisLabel}{character.}

\item{title}{character.}
}
\description{
Produces and prints out bar graph
}
\examples{
bargraphPlot(virginiaYears, virginiaData$vio, "Violence Crime Rate", "Violence Over Years")
}
\author{
Dhuha Manhil
}
