% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{boxPlot}
\alias{boxPlot}
\title{Produces a boxplot}
\usage{
boxPlot(dataSet, xaxis, yaxis, xaxisLabel, yaxisLabel, title)
}
\arguments{
\item{dataSet}{data}

\item{xaxis}{variable}

\item{yaxis}{variable}

\item{xaxisLabel}{character}

\item{yaxisLabel}{character}

\item{title}{character}
}
\description{
Produces and prints out a boxplot
}
\examples{
boxPlot(virginiaData, virginiaData$shall == 1, virginiaData$totalCrime, "Shall Law", "Total Crime",
"Total Crime vs Shall Law")
}
\author{
Dhuha Manhil
}
