% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{correlationFunction}
\alias{correlationFunction}
\title{Correlation between two variables}
\usage{
correlationFunction(x, y)
}
\arguments{
\item{x, y}{variable}
}
\description{
Calculates and prints out correlation
}
\examples{
correlationFunction(x = virginiaData$vio, y = virginiaData$mur)
}
\author{
Dhuha Manhil
}
