% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{histogramPlot}
\alias{histogramPlot}
\title{Produces a histogram plot}
\usage{
histogramPlot(dataSet, xaxis, xaxisLabel, yaxisLabel, title)
}
\arguments{
\item{dataSet}{data}

\item{xaxis}{variable}

\item{xaxisLabel}{character}

\item{yaxisLabel}{character}

\item{title}{character}
}
\description{
Produces and prints out histogram plot
}
\examples{
histogramPlot(virginiaData, virginiaData$totalCrime, "Total Crime Rate of Murders,
Violence, and Robbery", "Frequency", "Histogram of Total Crime Rates")
}
\author{
Dhuha Manhil
}
