% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{plottingFunction}
\alias{plottingFunction}
\title{Calculates linear regression}
\usage{
plottingFunction(x, y, a, b, colorofPoints, xaxis, yaxis)
}
\arguments{
\item{x}{variable.}

\item{y}{variable.}

\item{a}{variable.}

\item{b}{numeric}

\item{xaxis}{character.}

\item{yaxis}{character.}

\item{colorOfPoints}{character.}
}
\description{
Plots linear model with regression line
}
\examples{
plottingFunction(x = virginiaData$rob, y = virginiaData$avginc, a = virginiaData, b = 16)
}
\author{
Dhuha Manhil
}
