% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{readingInData}
\alias{readingInData}
\title{Reads in a .xlsx file}
\usage{
readingInData(file)
}
\arguments{
\item{dataFile}{variable}
}
\description{
Reads in a data file and prints out data
}
\examples{
readingInData(dataFile = "Guns.xlsx")
}
\author{
Dhuha Manhil
}
