% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhil-Dhuha-FinalProject-RCODE.R
\name{summarizeFunction}
\alias{summarizeFunction}
\title{Summarizes a data set}
\usage{
summarizeFunction(dataSet)
}
\arguments{
\item{dataSet}{variable}
}
\description{
Summarizes and prints out summary of data
}
\examples{
sumarizeFunction(dataSet = virginiaData)
}
\author{
Dhuha Manhil
}
